moved the namespace of post to models folder
