+++ Fakebook.Posts/Fakebook.Posts.DataAccess/PostsRepository.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Fakebook.Posts.Domain;
+using Fakebook.Posts.Domain.Models;
 
 namespace FakebookPosts.DataModel
 {
@@ -12,12 +13,12 @@ namespace FakebookPosts.DataModel
 
         public bool IsReadOnly => false;
 
-        public void Add(Fakebook.Posts.Domain.Post item)
+        public void Add(Fakebook.Posts.Domain.Models.Post item)
         {
             throw new System.NotImplementedException();
         }
 
-        public ValueTask<bool> AddAsync(Fakebook.Posts.Domain.Post post)
+        public ValueTask<Fakebook.Posts.Domain.Models.Post> AddAsync(Fakebook.Posts.Domain.Models.Post post)
         {
             throw new System.NotImplementedException();
         }
@@ -27,27 +28,27 @@ namespace FakebookPosts.DataModel
             throw new System.NotImplementedException();
         }
 
-        public bool Contains(Fakebook.Posts.Domain.Post item)
+        public bool Contains(Fakebook.Posts.Domain.Models.Post item)
         {
             throw new System.NotImplementedException();
         }
 
-        public void CopyTo(Fakebook.Posts.Domain.Post[] array, int arrayIndex)
+        public void CopyTo(Fakebook.Posts.Domain.Models.Post[] array, int arrayIndex)
         {
             throw new System.NotImplementedException();
         }
 
-        public IAsyncEnumerator<Fakebook.Posts.Domain.Post> GetAsyncEnumerator(CancellationToken cancellationToken = default)
+        public IAsyncEnumerator<Fakebook.Posts.Domain.Models.Post> GetAsyncEnumerator(CancellationToken cancellationToken = default)
         {
             throw new System.NotImplementedException();
         }
 
-        public IEnumerator<Fakebook.Posts.Domain.Post> GetEnumerator()
+        public IEnumerator<Fakebook.Posts.Domain.Models.Post> GetEnumerator()
         {
             throw new System.NotImplementedException();
         }
 
-        public bool Remove(Fakebook.Posts.Domain.Post item)
+        public bool Remove(Fakebook.Posts.Domain.Models.Post item)
         {
             throw new System.NotImplementedException();
         }
